wip: update designs step 1
