+++ README.md
@@ -4,10 +4,8 @@ Notes to the end of the line and back again.
 
 ## Attribution
 
-- [Design ideas](https://dribbble.com/shots/23237785-Glyph-Beer-15)
 - [Pacifico Font](https://fonts.google.com/specimen/Pacifico)
-- [CSS Doodle](https://css-doodle.com/)
+<!-- - [CSS Doodle](https://css-doodle.com/) -->
 - [ChatGPT](https://chat.openai.com/)
 - [Stable Diffusion](https://stablediffusionweb.com/#ai-image-generator)
 - [Logo](https://www.svgrepo.com/svg/528342/letter-opened)
-<!-- - [Pageflip](https://github.com/Nodlik/react-pageflip) -->
