+++ package.json
@@ -7,7 +7,6 @@
     "react-dom": "^18.2.0",
     "react-error-boundary": "^4.0.12",
     "react-list": "^0.8.17",
-    "react-pageflip": "^2.0.3",
     "react-query": "^3.39.3",
     "react-scripts": "5.0.1",
     "styled-components": "^6.1.6",
