+++ public/stories/credits.txt
@@ -1,0 +1,7 @@
+Well that was... something.
+
+And look at us, here, at the end of this little book.
+
+If you want more ravings then shoot me a message@letters-from-the-abyss.com.
+
+Thanks for reading!
