+++ public/stories/welcome.txt
@@ -1,0 +1,10 @@
+Traveller.
+
+How you arrived only you, and your stalkers, know.
+
+But while you are here, indulge me a moment and
+flip through my letters, written to myself
+from a dark place with hope for a brighter future.
+
+If you like any of them do not hesitate to share.
+See you all again at the finale ❤️
