+++ src/components/PoemCard.css
@@ -14,10 +14,28 @@
   flex-direction: column;
   transition: opacity 0.3s ease-in-out;
   box-shadow: 1px 1px 4px 4px color(--color-hl);
+  margin-top: 100vh;
 }
 
 .poem-card-content > *[class^="poem-card"] {
   background-color: var(--color-fg);
+}
+
+.poem-card-photo-wrap {
+  display: flex;
+  overflow: hidden;
+  align-items: center;
+  justify-content: center;
+  position: fixed;
+  top: calc(3em + 32px);
+  height: min(80vw, 80vh);
+  width: min(80vw, 80vh);
+  min-width: 480px;
+  min-height: 480px;
+  border-radius: 40px;
+  border-color: var(--color-fg);
+  border-width: 4px;
+  border-style: solid;
 }
 
 .poem-card-photo {
@@ -25,14 +43,13 @@
   align-items: center;
   justify-content: center;
   position: relative;
-  overflow: hidden;
   height: 100%;
   width: 100%;
 
   background-color: transparent;
-  background-position: top center;
+  background-position: center;
   background-repeat: no-repeat;
-  background-size: contain;
+  background-size: cover;
 
   &.collapsed {
     width: 0vw;
@@ -47,13 +64,6 @@
   /* height: calc(100% - 64px); */
   width: calc(100% - 64px);
 }
-
-/* .poem-card-text-sizer {
-  position: absolute;
-  top: 0;
-  left: 0;
-  visibility: hidden;
-} */
 
 .poem-card-timestamp {
   /* position: absolute;
