+++ src/utils/stories.json
@@ -23,6 +23,7 @@
       "bleeding-heart": "2022-12-26",
       "collapse": "2023-04-30",
       "collateral": "2023-05-18",
+      "credits": "9999-12-31",
       "dead-o-winter": "2014-12-08",
       "death-of-depression": "2020-09-07",
       "deicide": "2023-09-28",
@@ -46,7 +47,8 @@
       "the-woods": "2018-02-14",
       "umbrette": "2023-09-27",
       "walled-garden": "2015-11-26",
-      "wandering-heart": "2023-06-21"
+      "wandering-heart": "2023-06-21",
+      "welcome": "0000-01-01"
     }
   }
 }
